fix: phpstan improvements
