+++ core/Database/Concerns/Query/BuildsQuery.php
@@ -13,7 +13,6 @@ use Core\Database\Functions;
 use Core\Database\Having;
 use Core\Database\SelectCase;
 use Core\Database\Subquery;
-use Core\Exceptions\QueryError;
 use Core\Util\Arr;
 
 trait BuildsQuery
@@ -220,33 +219,6 @@ trait BuildsQuery
         }
 
         return Arr::implodeDeeply($query);
-    }
-
-    protected function prepareColumns(array $columns): string
-    {
-        $columns = array_map(function ($column) {
-            return match (true) {
-                $column instanceof Functions => (string) $column,
-                $column instanceof SelectCase => (string) $column,
-                $column instanceof Subquery => $this->resolveSubquery($column),
-                default => $column,
-            };
-        }, $columns);
-
-        return Arr::implodeDeeply($columns, ', ');
-    }
-
-    private function resolveSubquery(Subquery $subquery): string
-    {
-        [$dml, $arguments] = $subquery->toSql();
-
-        if (! str_contains($dml, 'LIMIT 1')) {
-            throw new QueryError('The subquery must be limited to one record');
-        }
-
-        $this->arguments = array_merge($this->arguments, $arguments);
-
-        return $dml;
     }
 
     private function prepareDataToInsert(array $data): void
