+++ core/Services/Manager.php
@@ -24,6 +24,8 @@ abstract class Manager
 
     protected function createDriver(string $driver): object
     {
-        # code...
+        $method = 'create' . ucfirst($driver) . 'Driver';
+
+        return $this->{$method}();
     }
 }
