+++ tests/Core/Database/Connections/ConnectionFactoryTest.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Core\Database\Connections;
+
+it('asserts truthy', function () {
+    expect(true)->toBeTruthy();
+});
